feat/favorites

+++ products/views.py
@@ -1,6 +1,6 @@
-from django.urls import reverse
 from django.contrib import messages
 from django.contrib.auth.decorators import login_required
+from django.views.decorators.csrf import csrf_exempt
 from django.http import JsonResponse
 from django.shortcuts import render, redirect, get_object_or_404
 from products.forms import ProductForm
@@ -10,10 +10,11 @@ from products.models import Product
 def main(request):
     dataset_products = Product.objects.filter(user__isnull=True)
     user_products = Product.objects.filter(user__isnull=False)
-
     products = dataset_products.union(user_products)
     
-    return render(request, 'database_products.html', {'products': products})
+    categories = Product.objects.values_list('category', flat=True).distinct()
+    
+    return render(request, 'database_products.html', {'products': products, 'categories': categories,})
 
 # Show all logged in merchant's products
 @login_required(login_url='/accounts/login')
@@ -33,6 +34,7 @@ def view_product(request, id):
     product = get_object_or_404(Product, pk=id)
     return render(request, 'view_product.html', {'product': product})
 
+@csrf_exempt
 @login_required(login_url='/accounts/login')
 def create_product(request):
     if request.method == "POST":
@@ -49,6 +51,7 @@ def create_product(request):
     context = {'form': form}
     return render(request, "create_product.html", context)
 
+@csrf_exempt
 @login_required(login_url='/accounts/login')
 def edit_product(request, id):
     product = get_object_or_404(Product, pk=id, user=request.user)
@@ -65,12 +68,15 @@ def edit_product(request, id):
     context = {'form': form, 'product': product}
     return render(request, "edit_product.html", context)
 
+@csrf_exempt
 @login_required(login_url='/accounts/login')
 def delete_product(request, id):
-    product = get_object_or_404(Product, pk=id, user=request.user)
-    product.delete()
-    messages.success(request, "Your product has been successfully deleted!")
-    return redirect('products:my_products')
+    if request.method == 'POST':
+        product = get_object_or_404(Product, pk=id)
+        product.delete()
+        return JsonResponse({"status": "success"})
+    else:
+        return JsonResponse({"status": "error", "message": "Invalid request method"}, status=400)
 
 # Show all categories
 def view_categories(request):
